backend/requirements.txt
fastapi
upstash-redis
aiohttp
python-dotenv
